refactor(tests): split tests/test_llm.py into tests/llms package
