+++ tests/llms/test_base.py
@@ -1,0 +1,2 @@
+"""Tests for src.llms.base."""
+from __future__ import annotations
