+++ tests/llms/test_composite.py
@@ -1,0 +1,17 @@
+"""Tests for src.llms.composite."""
+from __future__ import annotations
+
+from unittest.mock import MagicMock
+
+from src.llms.composite import _CompositeProvider
+
+
+def test_composite_provider_routes_image_to_image_llm():
+    text = MagicMock()
+    image = MagicMock()
+    image.generate_image.return_value = b"img"
+    composite = _CompositeProvider(text=text, image=image)
+    composite.chat(system="s", messages=[])
+    text.chat.assert_called_once()
+    composite.generate_image("x")
+    image.generate_image.assert_called_once_with("x")
