+++ tests/llms/test_openai_wire.py
@@ -1,0 +1,2 @@
+"""Tests for src.llms.openai_wire."""
+from __future__ import annotations
