--- tests/test_llm.py
@@ -1,651 +1,0 @@
-import json
-from unittest.mock import MagicMock, patch
-
-from src.llms import ToolCall, get_llm
-from src.llms.bedrock import BedrockProvider
-from src.llms.openai import OpenAIProvider
-from src.llms.xai import XAIProvider
-
-
-# --------------------------------------------------------------------------- #
-# OpenAI
-# --------------------------------------------------------------------------- #
-
-
-def _openai_completion(content="", tool_calls=None, finish="stop"):
-    choice = MagicMock()
-    choice.finish_reason = finish
-    choice.message.content = content
-    choice.message.tool_calls = tool_calls or []
-    completion = MagicMock()
-    completion.choices = [choice]
-    completion.usage.prompt_tokens = 10
-    completion.usage.completion_tokens = 20
-    return completion
-
-
-def _openai_tool_call(call_id, name, args_obj):
-    tc = MagicMock()
-    tc.id = call_id
-    tc.function.name = name
-    tc.function.arguments = json.dumps(args_obj)
-    return tc
-
-
-def test_openai_chat_parses_text():
-    provider = OpenAIProvider(model="gpt-4o-mini", image_model="gpt-image-1")
-    provider._client = MagicMock()
-    provider._client.chat.completions.create.return_value = _openai_completion(content="hello")
-    result = provider.chat(system="s", messages=[{"role": "user", "content": "hi"}])
-    assert result.content == "hello"
-    assert result.stop_reason == "end_turn"
-    assert result.tool_calls == []
-    assert result.token_usage == {"input": 10, "output": 20}
-
-
-def test_openai_legacy_model_uses_max_tokens_and_temperature():
-    provider = OpenAIProvider(model="gpt-4o-mini", image_model="gpt-image-1")
-    provider._client = MagicMock()
-    provider._client.chat.completions.create.return_value = _openai_completion(content="x")
-    provider.chat(system="s", messages=[])
-    kwargs = provider._client.chat.completions.create.call_args.kwargs
-    assert "max_tokens" in kwargs
-    assert "temperature" in kwargs
-    assert "max_completion_tokens" not in kwargs
-
-
-def test_openai_new_generation_uses_max_completion_tokens():
-    provider = OpenAIProvider(model="gpt-5.4", image_model="gpt-image-1")
-    provider._client = MagicMock()
-    provider._client.chat.completions.create.return_value = _openai_completion(content="x")
-    provider.chat(system="s", messages=[])
-    kwargs = provider._client.chat.completions.create.call_args.kwargs
-    assert "max_completion_tokens" in kwargs
-    assert "max_tokens" not in kwargs
-    assert "temperature" not in kwargs
-
-
-def test_openai_o1_model_uses_max_completion_tokens():
-    provider = OpenAIProvider(model="o1-mini", image_model="gpt-image-1")
-    provider._client = MagicMock()
-    provider._client.chat.completions.create.return_value = _openai_completion(content="x")
-    provider.chat(system="s", messages=[])
-    kwargs = provider._client.chat.completions.create.call_args.kwargs
-    assert "max_completion_tokens" in kwargs
-    assert "temperature" not in kwargs
-
-
-def _make_stream_chunk(*, content=None, tool_calls=None, finish=None, usage=None):
-    chunk = MagicMock()
-    if usage is not None:
-        chunk.usage = usage
-    else:
-        chunk.usage = None
-    if content is None and tool_calls is None and finish is None:
-        chunk.choices = []
-        return chunk
-    choice = MagicMock()
-    choice.finish_reason = finish
-    choice.delta.content = content
-    choice.delta.tool_calls = tool_calls
-    chunk.choices = [choice]
-    return chunk
-
-
-def _stream_tool_call(index, call_id=None, name=None, arguments=None):
-    tc = MagicMock()
-    tc.index = index
-    tc.id = call_id
-    fn = MagicMock()
-    fn.name = name
-    fn.arguments = arguments
-    tc.function = fn
-    return tc
-
-
-def test_openai_chat_streams_content_when_on_delta_given():
-    provider = OpenAIProvider(model="gpt-4o-mini", image_model="gpt-image-1")
-    provider._client = MagicMock()
-    chunks = [
-        _make_stream_chunk(content="hello"),
-        _make_stream_chunk(content=" world"),
-        _make_stream_chunk(finish="stop"),
-    ]
-    provider._client.chat.completions.create.return_value = iter(chunks)
-
-    received: list[str] = []
-    result = provider.chat(system="s", messages=[], on_delta=received.append)
-
-    assert received == ["hello", " world"]
-    assert result.content == "hello world"
-    assert result.stop_reason == "end_turn"
-    assert result.tool_calls == []
-    # must have requested streaming
-    kwargs = provider._client.chat.completions.create.call_args.kwargs
-    assert kwargs.get("stream") is True
-
-
-def test_openai_chat_stream_suppresses_content_after_tool_calls_begin():
-    provider = OpenAIProvider(model="gpt-4o-mini", image_model="gpt-image-1")
-    provider._client = MagicMock()
-    chunks = [
-        # tool_call begins first (no content yet)
-        _make_stream_chunk(tool_calls=[_stream_tool_call(0, call_id="c1", name="search_web")]),
-        _make_stream_chunk(tool_calls=[_stream_tool_call(0, arguments='{"query":')]),
-        _make_stream_chunk(tool_calls=[_stream_tool_call(0, arguments='"x"}')]),
-        # Model sometimes also emits commentary content after initiating a tool_call.
-        _make_stream_chunk(content="I'll search."),
-        _make_stream_chunk(finish="tool_calls"),
-    ]
-    provider._client.chat.completions.create.return_value = iter(chunks)
-
-    received: list[str] = []
-    result = provider.chat(
-        system="s",
-        messages=[],
-        tools=[{"name": "search_web", "description": "", "parameters": {"type": "object"}}],
-        on_delta=received.append,
-    )
-
-    # Content after tool_calls started must NOT reach on_delta.
-    assert received == []
-    assert result.tool_calls[0].name == "search_web"
-    assert result.tool_calls[0].arguments == {"query": "x"}
-    assert result.stop_reason == "tool_use"
-
-
-def test_openai_chat_stream_accumulates_usage_from_last_chunk():
-    provider = OpenAIProvider(model="gpt-4o-mini", image_model="gpt-image-1")
-    provider._client = MagicMock()
-    usage = MagicMock()
-    usage.prompt_tokens = 42
-    usage.completion_tokens = 7
-    chunks = [
-        _make_stream_chunk(content="hi"),
-        _make_stream_chunk(finish="stop", usage=usage),
-    ]
-    provider._client.chat.completions.create.return_value = iter(chunks)
-
-    result = provider.chat(system="s", messages=[], on_delta=lambda _: None)
-    assert result.token_usage == {"input": 42, "output": 7}
-
-
-def test_openai_translates_canonical_tool_calls():
-    """Canonical assistant tool_calls must be serialized to OpenAI's wire format."""
-    provider = OpenAIProvider(model="gpt-4o-mini", image_model="gpt-image-1")
-    provider._client = MagicMock()
-    provider._client.chat.completions.create.return_value = _openai_completion(content="done")
-
-    canonical = [
-        {"role": "user", "content": "ask"},
-        {
-            "role": "assistant",
-            "content": "",
-            "tool_calls": [{"id": "c1", "name": "search_web", "arguments": {"query": "q"}}],
-        },
-        {"role": "tool", "tool_call_id": "c1", "content": "{\"ok\": true}"},
-    ]
-    provider.chat(system="s", messages=canonical)
-    sent = provider._client.chat.completions.create.call_args.kwargs["messages"]
-    # system + 3 canonical = 4
-    assert len(sent) == 4
-    assistant = sent[2]
-    assert assistant["role"] == "assistant"
-    assert assistant["tool_calls"][0]["type"] == "function"
-    assert assistant["tool_calls"][0]["function"]["name"] == "search_web"
-    # arguments must be a JSON string, not a dict
-    assert isinstance(assistant["tool_calls"][0]["function"]["arguments"], str)
-    import json as _json
-
-    assert _json.loads(assistant["tool_calls"][0]["function"]["arguments"]) == {"query": "q"}
-
-
-def test_openai_chat_parses_tool_calls():
-    provider = OpenAIProvider(model="gpt-4o-mini", image_model="gpt-image-1")
-    provider._client = MagicMock()
-    tc = _openai_tool_call("call_1", "search_web", {"query": "aws"})
-    provider._client.chat.completions.create.return_value = _openai_completion(
-        tool_calls=[tc], finish="tool_calls"
-    )
-    result = provider.chat(system="s", messages=[], tools=[{"name": "search_web", "description": "", "parameters": {}}])
-    assert result.stop_reason == "tool_use"
-    assert len(result.tool_calls) == 1
-    assert result.tool_calls[0].name == "search_web"
-    assert result.tool_calls[0].arguments == {"query": "aws"}
-
-
-def test_openai_chat_handles_bad_tool_arguments():
-    provider = OpenAIProvider(model="gpt-4o-mini", image_model="gpt-image-1")
-    provider._client = MagicMock()
-    tc = MagicMock()
-    tc.id = "x"
-    tc.function.name = "search_web"
-    tc.function.arguments = "not json"
-    provider._client.chat.completions.create.return_value = _openai_completion(
-        tool_calls=[tc], finish="tool_calls"
-    )
-    result = provider.chat(system="s", messages=[], tools=[{"name": "search_web", "description": "", "parameters": {}}])
-    assert result.tool_calls[0].arguments == {}
-
-
-def test_openai_stream_chat_invokes_callback():
-    provider = OpenAIProvider(model="gpt-4o-mini", image_model="gpt-image-1")
-    provider._client = MagicMock()
-
-    def _chunk(text):
-        ch = MagicMock()
-        ch.choices[0].delta.content = text
-        return ch
-
-    provider._client.chat.completions.create.return_value = iter([_chunk("he"), _chunk("llo")])
-    seen = []
-    result = provider.stream_chat(system="s", messages=[], on_delta=seen.append)
-    assert result == "hello"
-    assert seen == ["he", "llo"]
-
-
-# --------------------------------------------------------------------------- #
-# Bedrock — Claude
-# --------------------------------------------------------------------------- #
-
-
-def _bedrock_response(payload: dict):
-    body = MagicMock()
-    body.read.return_value = json.dumps(payload).encode()
-    return {"body": body}
-
-
-def test_bedrock_inference_profile_routes_to_claude():
-    """us.anthropic.claude-* inference profile IDs must still hit the Claude
-    path (tool_use, messages API), not the unknown-family fallback."""
-    provider = BedrockProvider(
-        model="us.anthropic.claude-opus-4-6-v1",
-        image_model="amazon.nova-canvas-v1:0",
-        region="us-east-1",
-    )
-    provider._client = MagicMock()
-    provider._client.invoke_model.return_value = _bedrock_response(
-        {
-            "content": [{"type": "text", "text": "hi"}],
-            "stop_reason": "end_turn",
-            "usage": {"input_tokens": 1, "output_tokens": 2},
-        }
-    )
-    tools = [{"name": "search_web", "description": "", "parameters": {"type": "object"}}]
-    result = provider.chat(system="s", messages=[], tools=tools)
-    assert result.content == "hi"
-    # Claude body carries tools (not Nova's toolConfig)
-    body = provider._client.invoke_model.call_args.kwargs["body"]
-    import json as _json
-
-    parsed = _json.loads(body)
-    assert "tools" in parsed  # routed into _claude_chat, not fallback
-    assert parsed["tools"][0]["name"] == "search_web"
-
-
-def test_bedrock_inference_profile_image_routing():
-    """global./us. prefixed image model IDs still reach the Titan/Nova body."""
-    provider = BedrockProvider(
-        model="us.anthropic.claude-opus-4-6-v1",
-        image_model="us.amazon.nova-canvas-v1:0",
-        region="us-east-1",
-    )
-    body = provider._build_image_body("cat")
-    assert body["taskType"] == "TEXT_IMAGE"
-
-
-def test_bedrock_claude_chat_text():
-    provider = BedrockProvider(
-        model="anthropic.claude-3-5-sonnet-20240620-v1:0",
-        image_model="amazon.titan-image-generator-v1",
-        region="us-east-1",
-    )
-    provider._client = MagicMock()
-    provider._client.invoke_model.return_value = _bedrock_response(
-        {
-            "content": [{"type": "text", "text": "안녕"}],
-            "stop_reason": "end_turn",
-            "usage": {"input_tokens": 5, "output_tokens": 7},
-        }
-    )
-    result = provider.chat(system="s", messages=[{"role": "user", "content": "hi"}])
-    assert result.content == "안녕"
-    assert result.stop_reason == "end_turn"
-    assert result.token_usage == {"input": 5, "output": 7}
-
-
-def test_bedrock_claude_chat_with_tool_use():
-    provider = BedrockProvider(
-        model="anthropic.claude-3-5-sonnet-20240620-v1:0",
-        image_model="amazon.titan-image-generator-v1",
-        region="us-east-1",
-    )
-    provider._client = MagicMock()
-    provider._client.invoke_model.return_value = _bedrock_response(
-        {
-            "content": [
-                {"type": "text", "text": "I'll search."},
-                {"type": "tool_use", "id": "tu_1", "name": "search_web", "input": {"query": "x"}},
-            ],
-            "stop_reason": "tool_use",
-            "usage": {"input_tokens": 3, "output_tokens": 4},
-        }
-    )
-    tools = [{"name": "search_web", "description": "", "parameters": {"type": "object"}}]
-    result = provider.chat(system="s", messages=[], tools=tools)
-    assert result.stop_reason == "tool_use"
-    assert result.tool_calls[0].name == "search_web"
-    assert result.tool_calls[0].arguments == {"query": "x"}
-
-
-def test_bedrock_message_translation_tool_role():
-    messages = [
-        {"role": "user", "content": "ask"},
-        {"role": "assistant", "content": "", "tool_calls": [{"id": "t1", "name": "foo", "arguments": {"a": 1}}]},
-        {"role": "tool", "tool_call_id": "t1", "content": "{\"ok\":true}"},
-    ]
-    translated = BedrockProvider._to_anthropic_messages(messages)
-    assert translated[0] == {"role": "user", "content": "ask"}
-    assert translated[1]["role"] == "assistant"
-    assert translated[1]["content"][0]["type"] == "tool_use"
-    assert translated[1]["content"][0]["name"] == "foo"
-    assert translated[2]["role"] == "user"
-    assert translated[2]["content"][0]["type"] == "tool_result"
-
-
-# --------------------------------------------------------------------------- #
-# Bedrock — Nova
-# --------------------------------------------------------------------------- #
-
-
-def test_bedrock_nova_chat_text():
-    provider = BedrockProvider(model="amazon.nova-pro-v1:0", image_model="amazon.nova-canvas-v1:0", region="us-east-1")
-    provider._client = MagicMock()
-    provider._client.converse.return_value = {
-        "output": {"message": {"content": [{"text": "hi"}]}},
-        "stopReason": "end_turn",
-        "usage": {"inputTokens": 1, "outputTokens": 2},
-    }
-    result = provider.chat(system="s", messages=[{"role": "user", "content": "hi"}])
-    assert result.content == "hi"
-    assert result.stop_reason == "end_turn"
-    assert result.token_usage == {"input": 1, "output": 2}
-
-
-def test_bedrock_nova_tool_use():
-    provider = BedrockProvider(model="amazon.nova-pro-v1:0", image_model="amazon.nova-canvas-v1:0", region="us-east-1")
-    provider._client = MagicMock()
-    provider._client.converse.return_value = {
-        "output": {
-            "message": {
-                "content": [
-                    {"text": "let me search"},
-                    {"toolUse": {"toolUseId": "tu1", "name": "search_web", "input": {"query": "q"}}},
-                ]
-            }
-        },
-        "stopReason": "tool_use",
-        "usage": {"inputTokens": 2, "outputTokens": 3},
-    }
-    result = provider.chat(
-        system="s",
-        messages=[],
-        tools=[{"name": "search_web", "description": "", "parameters": {"type": "object"}}],
-    )
-    assert result.stop_reason == "tool_use"
-    assert result.tool_calls[0].name == "search_web"
-
-
-# --------------------------------------------------------------------------- #
-# Image model adapter
-# --------------------------------------------------------------------------- #
-
-
-def test_build_image_body_titan():
-    provider = BedrockProvider(
-        model="anthropic.claude-3-5-sonnet-20240620-v1:0",
-        image_model="amazon.titan-image-generator-v1",
-        region="us-east-1",
-    )
-    body = provider._build_image_body("a cat")
-    assert body["taskType"] == "TEXT_IMAGE"
-
-
-def test_build_image_body_stability():
-    provider = BedrockProvider(
-        model="anthropic.claude-3-5-sonnet-20240620-v1:0",
-        image_model="stability.stable-diffusion-xl-v1",
-        region="us-east-1",
-    )
-    body = provider._build_image_body("a cat")
-    assert body["text_prompts"][0]["text"] == "a cat"
-
-
-def test_build_image_body_unknown_raises():
-    provider = BedrockProvider(
-        model="anthropic.claude-3-5-sonnet-20240620-v1:0",
-        image_model="mystery.v1",
-        region="us-east-1",
-    )
-    import pytest
-
-    with pytest.raises(ValueError):
-        provider._build_image_body("x")
-
-
-def test_bedrock_describe_image_returns_text():
-    provider = BedrockProvider(
-        model="anthropic.claude-3-5-sonnet-20240620-v1:0",
-        image_model="amazon.titan-image-generator-v1",
-        region="us-east-1",
-    )
-    provider._client = MagicMock()
-    provider._client.invoke_model.return_value = _bedrock_response(
-        {"content": [{"type": "text", "text": "a cat"}]}
-    )
-    out = provider.describe_image(b"fake", "image/png")
-    assert out == "a cat"
-
-
-def test_bedrock_generate_image_titan_returns_bytes():
-    import base64 as _b64
-
-    provider = BedrockProvider(
-        model="anthropic.claude-3-5-sonnet-20240620-v1:0",
-        image_model="amazon.titan-image-generator-v1",
-        region="us-east-1",
-    )
-    provider._client = MagicMock()
-    provider._client.invoke_model.return_value = _bedrock_response(
-        {"images": [_b64.b64encode(b"imgdata").decode()]}
-    )
-    assert provider.generate_image("cat") == b"imgdata"
-
-
-def test_bedrock_generate_image_stability_returns_bytes():
-    import base64 as _b64
-
-    provider = BedrockProvider(
-        model="anthropic.claude-3-5-sonnet-20240620-v1:0",
-        image_model="stability.stable-diffusion-xl-v1",
-        region="us-east-1",
-    )
-    provider._client = MagicMock()
-    provider._client.invoke_model.return_value = _bedrock_response(
-        {"artifacts": [{"base64": _b64.b64encode(b"xyz").decode()}]}
-    )
-    assert provider.generate_image("cat") == b"xyz"
-
-
-def test_openai_describe_image_uses_vision_format():
-    provider = OpenAIProvider(model="gpt-4o-mini", image_model="gpt-image-1")
-    provider._client = MagicMock()
-    provider._client.chat.completions.create.return_value = _openai_completion(content="it's a cat")
-    out = provider.describe_image(b"\x89PNG", "image/png")
-    assert out == "it's a cat"
-    args = provider._client.chat.completions.create.call_args.kwargs
-    assert args["messages"][0]["content"][1]["type"] == "image_url"
-
-
-def test_openai_generate_image_decodes_b64():
-    import base64 as _b64
-
-    provider = OpenAIProvider(model="gpt-4o-mini", image_model="gpt-image-1")
-    provider._client = MagicMock()
-    response = MagicMock()
-    response.data = [MagicMock(b64_json=_b64.b64encode(b"hello").decode())]
-    provider._client.images.generate.return_value = response
-    assert provider.generate_image("cat") == b"hello"
-    kwargs = provider._client.images.generate.call_args.kwargs
-    # gpt-image-1 must NOT send response_format (API rejects it)
-    assert "response_format" not in kwargs
-
-
-def test_openai_generate_image_dalle_sends_response_format():
-    import base64 as _b64
-
-    provider = OpenAIProvider(model="gpt-4o-mini", image_model="dall-e-3")
-    provider._client = MagicMock()
-    response = MagicMock()
-    response.data = [MagicMock(b64_json=_b64.b64encode(b"ok").decode())]
-    provider._client.images.generate.return_value = response
-    provider.generate_image("cat")
-    kwargs = provider._client.images.generate.call_args.kwargs
-    assert kwargs["response_format"] == "b64_json"
-
-
-def test_composite_provider_routes_image_to_image_llm():
-    from src.llms.composite import _CompositeProvider
-
-    text = MagicMock()
-    image = MagicMock()
-    image.generate_image.return_value = b"img"
-    composite = _CompositeProvider(text=text, image=image)
-    composite.chat(system="s", messages=[])
-    text.chat.assert_called_once()
-    composite.generate_image("x")
-    image.generate_image.assert_called_once_with("x")
-
-
-# --------------------------------------------------------------------------- #
-# xAI (Grok)
-# --------------------------------------------------------------------------- #
-
-
-def test_xai_provider_uses_xai_base_url_and_api_key():
-    """XAIProvider must instantiate OpenAI client with the xAI base URL and
-    the explicit api_key, so traffic goes to api.x.ai rather than OpenAI."""
-    provider = XAIProvider(
-        model="grok-4-1-fast-reasoning",
-        image_model="grok-imagine-image",
-        api_key="xai-test",
-    )
-    with patch("openai.OpenAI") as openai_ctor:
-        openai_ctor.return_value = MagicMock()
-        provider._get_client()
-    kwargs = openai_ctor.call_args.kwargs
-    assert kwargs.get("base_url") == "https://api.x.ai/v1"
-    assert kwargs.get("api_key") == "xai-test"
-
-
-def test_xai_chat_parses_tool_calls():
-    """Grok returns the same wire shape as OpenAI for tool calls; the
-    shared parser must turn them into ToolCall objects."""
-    provider = XAIProvider(model="grok-4-1-fast-reasoning", image_model="grok-imagine-image", api_key="x")
-    provider._client = MagicMock()
-    tc = _openai_tool_call("call_g1", "search_web", {"query": "xai"})
-    provider._client.chat.completions.create.return_value = _openai_completion(
-        tool_calls=[tc], finish="tool_calls"
-    )
-    result = provider.chat(
-        system="s",
-        messages=[],
-        tools=[{"name": "search_web", "description": "", "parameters": {"type": "object"}}],
-    )
-    assert result.stop_reason == "tool_use"
-    assert result.tool_calls[0].name == "search_web"
-    assert result.tool_calls[0].arguments == {"query": "xai"}
-
-
-def test_xai_chat_uses_legacy_max_tokens_always():
-    """All current grok chat models accept max_tokens + temperature;
-    XAIProvider must not switch to max_completion_tokens (OpenAI-only split)."""
-    provider = XAIProvider(model="grok-4.20-0309-reasoning", image_model="grok-imagine-image", api_key="x")
-    provider._client = MagicMock()
-    provider._client.chat.completions.create.return_value = _openai_completion(content="hi")
-    provider.chat(system="s", messages=[])
-    kwargs = provider._client.chat.completions.create.call_args.kwargs
-    assert "max_tokens" in kwargs
-    assert "temperature" in kwargs
-    assert "max_completion_tokens" not in kwargs
-
-
-def test_xai_generate_image_skips_size_and_requests_b64():
-    """xAI images.generate rejects `size` (uses aspect_ratio/resolution).
-    We must omit it and explicitly ask for b64_json so we can decode bytes
-    into files_upload_v2."""
-    import base64 as _b64
-
-    provider = XAIProvider(model="grok-4-1-fast-reasoning", image_model="grok-imagine-image", api_key="x")
-    provider._client = MagicMock()
-    response = MagicMock()
-    response.data = [MagicMock(b64_json=_b64.b64encode(b"xai-bytes").decode())]
-    provider._client.images.generate.return_value = response
-
-    assert provider.generate_image("a cat") == b"xai-bytes"
-    kwargs = provider._client.images.generate.call_args.kwargs
-    assert kwargs["model"] == "grok-imagine-image"
-    assert kwargs["prompt"] == "a cat"
-    assert kwargs["response_format"] == "b64_json"
-    assert "size" not in kwargs  # xAI rejects this
-
-
-def test_xai_stream_chat_emits_deltas():
-    provider = XAIProvider(model="grok-4-1-fast-reasoning", image_model="grok-imagine-image", api_key="x")
-    provider._client = MagicMock()
-
-    def _chunk(text):
-        ch = MagicMock()
-        ch.choices[0].delta.content = text
-        return ch
-
-    provider._client.chat.completions.create.return_value = iter([_chunk("gr"), _chunk("ok")])
-    seen: list[str] = []
-    result = provider.stream_chat(system="s", messages=[], on_delta=seen.append)
-    assert result == "grok"
-    assert seen == ["gr", "ok"]
-
-
-# --------------------------------------------------------------------------- #
-# Factory
-# --------------------------------------------------------------------------- #
-
-
-def test_get_llm_builds_xai_provider():
-    provider = get_llm(
-        provider="xai",
-        model="grok-4-1-fast-reasoning",
-        image_provider="xai",
-        image_model="grok-imagine-image",
-        region="us-east-1",
-        api_keys={"xai": "xai-secret"},
-    )
-    assert isinstance(provider, XAIProvider)
-    assert provider._api_key == "xai-secret"
-
-
-def test_get_llm_composite_xai_text_openai_image():
-    """Mixed-provider setups still work through _CompositeProvider."""
-    from src.llms.composite import _CompositeProvider
-
-    provider = get_llm(
-        provider="xai",
-        model="grok-4-1-fast-reasoning",
-        image_provider="openai",
-        image_model="gpt-image-1",
-        region="us-east-1",
-        api_keys={"xai": "xai-key"},
-    )
-    assert isinstance(provider, _CompositeProvider)
-    assert isinstance(provider.text, XAIProvider)
-    assert isinstance(provider.image, OpenAIProvider)
